dom/html/test/test_bug1279218.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>Test for Bug 1279218</title>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
    <script type="text/javascript">
     function load() {
       let applets = document.applets;
       is(applets.length, 0, "Applet list length should be 0, even with applet tag in body");
       SimpleTest.finish();
     }

     window.onload=load;

     SimpleTest.waitForExplicitFinish();
    </script>
  </head>
  <body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1279218">Mozilla Bug 1279218</a>
    <applet id="applet-test"></applet>
  </body>
</html>
